LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-nested-async-expected.txt


FAIL Nested shadow roots independently resolve async adoptedStyleSheets. assert_equals: Outer root: expected 1 placeholder(s). expected 1 but got 0
